+++ packages/plugin-packages/mashroom-portal-sandbox-app/src/js/load_portal_app.ts
@@ -1,5 +1,4 @@
 
-import {isES6Module} from '@mashroom/mashroom-utils/lib/file-type-utils';
 import getClientServices from './client_services';
 
 import type {
@@ -17,7 +16,7 @@ const loadJs = (path: string): Promise<void> => {
     return new Promise((resolve, reject) => {
         console.info('Loading JS resource: ', path);
         const scriptElem = document.createElement('script');
-        if (isES6Module(path)) {
+        if (path.indexOf('.mjs') !== -1) {
             scriptElem.type = 'module';
         }
         scriptElem.src = path;