+++ packages/plugin-packages/mashroom-portal/src/frontend/portal-client/js/ResourceManager.ts
@@ -1,5 +1,4 @@
 
-import {isES6Module} from '@mashroom/mashroom-utils/lib/file-type-utils';
 import {WINDOW_VAR_PORTAL_INLINED_STYLE_APPS} from '../../../backend/constants';
 
 import type {LoadedPortalAppInternal} from './MashroomPortalAppServiceImpl';
@@ -57,7 +56,7 @@ export default class ResourceManager {
             };
 
             scriptElem.src = path;
-            if (isES6Module(path)) {
+            if (path.indexOf('.mjs') !== -1) {
                 scriptElem.type = 'module';
             }
             scriptElem.addEventListener('error', (error: any) => {