+++ packages/plugin-packages/mashroom-portal/src/frontend/portal-client/js/RestError.ts
@@ -1,5 +1,22 @@
 
-import {mergeStackTrace} from '@mashroom/mashroom-utils/lib/error-utils';
+const mergeStackTrace = (stackTrace: string, rootCauseStackTrace: string | null | undefined): string => {
+    if (!rootCauseStackTrace) {
+        return stackTrace;
+    }
+
+    const entriesToMerge = stackTrace.split('\n');
+    const baseEntries = rootCauseStackTrace.split('\n');
+
+    const newEntries: Array<string> = [];
+    entriesToMerge.forEach((entry) => {
+        if (baseEntries.includes(entry)) {
+            return;
+        }
+        newEntries.push(entry);
+    });
+
+    return [...newEntries, ...baseEntries].join('\n');
+};
 
 export class RestError extends Error {
     constructor(private statusCode: number, message: string, rootCauseStack?: string) {